# Conflicts:
#	frontend/lib/blocs/diet_generation/meal_recipe_bloc.dart
#	frontend/lib/config/endpoints.dart
#	frontend/lib/repository/api_client.dart
#	frontend/lib/repository/diet_generation/diet_prediction_repository.dart
#	frontend/test/mocks/mocks.mocks.dart
#	frontend/test/views/diet_generation/meal_recipe_screen_test.dart

+++ backend/user_details/schemas.py
@@ -74,9 +74,9 @@ class UserDetailsUpdate(DietGoalValidationMixin, DateOfBirthValidationMixin, Flo
 
 
 class PredictedMacros(BaseModel):
-    protein: int
-    fat: int
-    carbs: int
+    protein: float
+    fat: float
+    carbs: float
 
 
 class PredictedCalories(BaseModel):
